[nemo-qml-plugin-thumbnailer] Fix borders on thumbnails.
Centering a 180x180 rectangle around a point at 90x90 gives a rectangle
with the top left at (1,1) not (0,0) so the crop rectangle ends up
slightly outside the bounds of the image.  Biasing the width and height by 1
when dividing to get the center gives a 1x1 negative offset necessary to
get the correct center for even sizes without overcompensating for odd
widths.