[thumbnailer] avoid dangling pointer if previous request is removed from cache. Fixes JB#51775
Keeping a pointer to previousRequest that is already removed from cache
caused an application to crash due to double free (happened in jolla-camera
on Inoi T8 aarch64).

Improve cache accounting in NemoThumbnailLoader by storing the
value added to m_totalCost when a request is completed and subtracting
it when the request is removed from cache. The problem was that QImage
in request->pixmap can be emptied before request itself is deleted,
leading to associated value not getting subtracted from m_totalCost.